+++ seqr/views/apis/anvil_workspace_api.py
@@ -184,6 +184,7 @@ def create_project_from_workspace(request, namespace, name):
         'workspace_name': name,
         'mme_primary_data_owner': request.user.get_full_name(),
         'mme_contact_url': 'mailto:{}'.format(request.user.email),
+        'vlm_contact_email': request.user.email,
     }
 
     project = create_model_from_json(Project, project_args, user=request.user)